pyproject.toml
[build-system]
requires = ["setuptools>=80.10.2", "wheel>=0.46.3"]
build-backend = "setuptools.build_meta"

[project]
name = "ragonometrics"
version = "0.1.0"
description = "Simple RAG pipeline for economics papers"
requires-python = ">=3.11"
dependencies = [
  "openai>=2.17.0",
  "streamlit>=1.54.0",
  "requests>=2.32.5",
  "networkx>=3.6.1",
  "plotly>=6.5.2",
  "faiss-cpu>=1.13.2",
  "redis>=7.1.0",
  "rq>=2.6.1",
  "backoff>=2.2.1",
  "psycopg2-binary>=2.9.11",
  "rank_bm25>=0.2.2",
  "numpy>=2.4.2",
  "jiter>=0.13.0",
  "pydantic-core>=2.41.5",
  "pydantic>=2.12.5",
  "tomli>=2.4.0",
]
  

[project.optional-dependencies]
# Developer/test dependencies
dev = [
  "pytest>=9.0.2",
]

# Optional OCR/image dependencies (install only if you need OCR fallback)
ocr = [
  "pillow>=12.1.0",
  "pdf2image>=1.17.0",
  "pytesseract>=0.3.13",
]


[project.scripts]
ragonometrics = "ragonometrics.entrypoints:main"
ragonometrics-cli = "ragonometrics.cli:main"

[tool.setuptools]
packages = ["ragonometrics"]

[tool.setuptools.package-dir]
ragonometrics = "ragonometrics"

